requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.25
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
python-multipart==0.0.6
alembic==1.12.1
pydantic_settings==2.12.0
